// contracts/tasks/helpers/types.ts
export interface ITreeMergeParams {
  /**
   * Ethers signer
   */
  deployer: HardhatEthersSigner;

  /**
   * Poll contract
   */
  pollContract: Poll;

  /**
   * Message AccQueue contract
   */
  messageAccQueueContract: AccQueue;
}